# JIT: optimize out calls to term compare