Continuation of #1770 

Avoid calls to TERM_COMPARE primitive for comparison with literals in OP_IS_EXACT and OP_IS_NOT_EXACT.

Current benchmark:
|test                 |JIT disabled (c06edb2bca8dacc8c17661575b39bcfc2c5922ea)|HEAD~1, JIT enabled (c06edb2bca8dacc8c17661575b39bcfc2c5922ea)|This PR (9b14b166899f3248189b71c6c157c6ae110df6c4) |
|--------------|---------------- |------------------------|--------|
|erlang-tests   | 11.45s                  |  10.51s                            | 9.95s  |
|benchmark (entire run)                      | 17.03                 |  15.71 (incl. compilation)           | 15.35s
|benchmark : pingpong_speed_test | 3.59s                 |  3.20s                            | 3.23s |
|benchmark : prime_speed_test       |  0.26s                            | 0.21s           | 0.17s |
|benchmark : sudoku_solution_test |  0.26s                            | 1.35s (incl. compilation of sudoku_grid in 1.1s)         | 1.34s |
|benchmark : sudoku_puzzle_test   |    10.51s                            | 9.23s          | 9.18s |


These changes are made under both the "Apache 2.0" and the "GNU Lesser General
Public License 2.1 or later" license terms (dual license).

SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later